@@ -2,6 +2,7 @@ package edu.opjms.global
 
 import javafx.css.PseudoClass
 import javafx.scene.shape.SVGPath
+import javafx.util.Duration
 
 val plusIcon
     get() = svgPath("M24,13 L15,13 L15,4 C15,3.447 14.553,3 14,3 C13.447,3 13,3.447 13,4 L13,13 L4,13 C3.447,13 3,13.447 3,14 C3,14.553 3.447,15 4,15 L13,15 L13,24 C13,24.553 13.447,25 14,25 C14.553,25 15,24.553 15,24 L15,15 L24,15 C24.553,15 25,14.553 25,14 C25,13.447 24.553,13 24,13")
@@ -12,6 +13,38 @@ val plusIcon24
 val deleteIcon24
     get() = svgPath("M12 1.75C13.733 1.75 15.1492 3.10645 15.2449 4.81558L15.25 5H20.5C20.9142 5 21.25 5.33579 21.25 5.75C21.25 6.1297 20.9678 6.44349 20.6018 6.49315L20.5 6.5H19.704L18.4239 19.5192C18.2912 20.8683 17.1984 21.91 15.8626 21.9945L15.6871 22H8.31293C6.95734 22 5.81365 21.0145 5.59883 19.6934L5.57614 19.5192L4.295 6.5H3.5C3.1203 6.5 2.80651 6.21785 2.75685 5.85177L2.75 5.75C2.75 5.3703 3.03215 5.05651 3.39823 5.00685L3.5 5H8.75C8.75 3.20507 10.2051 1.75 12 1.75ZM18.197 6.5H5.802L7.06893 19.3724C7.12768 19.9696 7.60033 20.4343 8.18585 20.4936L8.31293 20.5H15.6871C16.2872 20.5 16.7959 20.0751 16.9123 19.4982L16.9311 19.3724L18.197 6.5ZM13.75 9.25C14.1297 9.25 14.4435 9.53215 14.4932 9.89823L14.5 10V17C14.5 17.4142 14.1642 17.75 13.75 17.75C13.3703 17.75 13.0565 17.4678 13.0068 17.1018L13 17V10C13 9.58579 13.3358 9.25 13.75 9.25ZM10.25 9.25C10.6297 9.25 10.9435 9.53215 10.9932 9.89823L11 10V17C11 17.4142 10.6642 17.75 10.25 17.75C9.8703 17.75 9.55651 17.4678 9.50685 17.1018L9.5 17V10C9.5 9.58579 9.83579 9.25 10.25 9.25ZM12 3.25C11.0818 3.25 10.3288 3.95711 10.2558 4.85647L10.25 5H13.75C13.75 4.0335 12.9665 3.25 12 3.25Z")
 
-val ERROR_CLASS = PseudoClass.getPseudoClass("error")
+val deleteIcon20
+    get() = svgPath("M11.5 4C11.5 3.17157 10.8284 2.5 10 2.5C9.17157 2.5 8.5 3.17157 8.5 4H7.5C7.5 2.61929 8.61929 1.5 10 1.5C11.3807 1.5 12.5 2.61929 12.5 4H17C17.2761 4 17.5 4.22386 17.5 4.5C17.5 4.77614 17.2761 5 17 5H16.446L15.1499 16.2292C15.0335 17.2384 14.179 18 13.1631 18H6.83688C5.821 18 4.9665 17.2384 4.85006 16.2292L3.553 5H3C2.75454 5 2.55039 4.82312 2.50806 4.58988L2.5 4.5C2.5 4.22386 2.72386 4 3 4H11.5ZM15.438 5H4.561L5.84347 16.1146C5.90169 16.6192 6.32894 17 6.83688 17H13.1631C13.6711 17 14.0983 16.6192 14.1565 16.1146L15.438 5ZM8.5 7.5C8.74546 7.5 8.94961 7.65477 8.99194 7.85886L9 7.9375V14.0625C9 14.3041 8.77614 14.5 8.5 14.5C8.25454 14.5 8.05039 14.3452 8.00806 14.1411L8 14.0625V7.9375C8 7.69588 8.22386 7.5 8.5 7.5ZM11.5 7.5C11.7455 7.5 11.9496 7.65477 11.9919 7.85886L12 7.9375V14.0625C12 14.3041 11.7761 14.5 11.5 14.5C11.2545 14.5 11.0504 14.3452 11.0081 14.1411L11 14.0625V7.9375C11 7.69588 11.2239 7.5 11.5 7.5Z")
 
-private fun svgPath(content: String) = SVGPath().also { it.content = content }
+val copyIcon24
+    get() = svgPath("M5.5028 4.62704L5.5 6.75V17.2542C5.5 19.0491 6.95507 20.5042 8.75 20.5042L17.3663 20.5045C17.0573 21.3782 16.224 22.0042 15.2444 22.0042H8.75C6.12665 22.0042 4 19.8776 4 17.2542V6.75C4 5.76929 4.62745 4.93512 5.5028 4.62704ZM17.75 2C18.9926 2 20 3.00736 20 4.25V17.25C20 18.4926 18.9926 19.5 17.75 19.5H8.75C7.50736 19.5 6.5 18.4926 6.5 17.25V4.25C6.5 3.00736 7.50736 2 8.75 2H17.75ZM17.75 3.5H8.75C8.33579 3.5 8 3.83579 8 4.25V17.25C8 17.6642 8.33579 18 8.75 18H17.75C18.1642 18 18.5 17.6642 18.5 17.25V4.25C18.5 3.83579 18.1642 3.5 17.75 3.5Z")
+
+val fileIcon24
+    get() = svgPath("M18.5 20C18.5 20.275 18.276 20.5 18 20.5H6C5.724 20.5 5.5 20.275 5.5 20V4C5.5 3.725 5.724 3.5 6 3.5H12V8C12 9.104 12.896 10 14 10H18.5V20ZM13.5 4.621L17.378 8.5H14C13.724 8.5 13.5 8.275 13.5 8V4.621ZM19.414 8.414L13.585 2.586C13.559 2.56 13.527 2.54 13.5 2.516C13.429 2.452 13.359 2.389 13.281 2.336C13.241 2.309 13.195 2.291 13.153 2.268C13.082 2.228 13.012 2.184 12.937 2.152C12.74 2.07 12.528 2.029 12.313 2.014C12.266 2.011 12.22 2 12.172 2H12.171H12H6C4.896 2 4 2.896 4 4V20C4 21.104 4.896 22 6 22H18C19.104 22 20 21.104 20 20V10V9.828C20 9.298 19.789 8.789 19.414 8.414Z")
+
+val arrowUp24
+    get() = svgPath("M4.28401 10.2959C3.89639 10.6893 3.90108 11.3225 4.29449 11.7101C4.68789 12.0977 5.32104 12.093 5.70866 11.6996L11 6.32931V20.0004C11 20.5527 11.4477 21.0004 12 21.0004C12.5523 21.0004 13 20.5527 13 20.0004V6.33579L18.2849 11.6996C18.6726 12.093 19.3057 12.0977 19.6991 11.7101C20.0925 11.3225 20.0972 10.6893 19.7096 10.2959L12.8872 3.37171C12.3976 2.8748 11.596 2.87479 11.1064 3.37171L4.28401 10.2959Z")
+
+val arrowDown24
+    get() = svgPath("M19.7159 13.7045C20.1036 13.3111 20.0989 12.678 19.7055 12.2903C19.3121 11.9027 18.6789 11.9074 18.2913 12.3008L13 17.6711V4C13 3.44771 12.5522 3 12 3C11.4477 3 11 3.44772 11 4V17.6646L5.71501 12.3008C5.32739 11.9074 4.69424 11.9027 4.30083 12.2903C3.90743 12.678 3.90274 13.3111 4.29036 13.7045L11.1127 20.6287C11.6024 21.1256 12.4039 21.1256 12.8936 20.6287L19.7159 13.7045Z")
+
+val deleteIcon
+    get() = svgPath("M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zm2.46-7.12l1.41-1.41L12 12.59l2.12-2.12 1.41 1.41L13.41 14l2.12 2.12-1.41 1.41L12 15.41l-2.12 2.12-1.41-1.41L10.59 14l-2.13-2.12zM15.5 4l-1-1h-5l-1 1H5v2h14V4z")
+
+val editIcon20
+    get() = svgPath("M13.245 2.81706C14.3318 1.73025 16.0939 1.73025 17.1807 2.81706C18.2222 3.85858 18.2656 5.52026 17.3109 6.61346L17.1807 6.75273L7.57485 16.3586C7.36976 16.5636 7.12302 16.7212 6.85215 16.821L6.68687 16.8739L2.6319 17.9798C2.28531 18.0743 1.96524 17.7857 2.00279 17.4452L2.01796 17.3658L3.12386 13.3109C3.20017 13.031 3.33624 12.7718 3.52191 12.5508L3.63917 12.4229L13.245 2.81706ZM12.3848 5.09195L4.34628 13.13C4.25399 13.2223 4.18096 13.3314 4.13089 13.4511L4.08862 13.574L3.21199 16.7847L6.42375 15.9091C6.5077 15.8862 6.58793 15.8526 6.66256 15.8093L6.77006 15.7372L6.86774 15.6515L14.9058 7.61295L12.3848 5.09195ZM16.4736 3.52417C15.816 2.86656 14.7725 2.83003 14.072 3.41457L13.9521 3.52417L13.0918 4.38495L15.6128 6.90595L16.4736 6.04563C17.1312 5.38803 17.1677 4.34455 16.5832 3.64407L16.4736 3.52417Z")
+
+val checkmark20
+    get() = svgPath("M3.37371 10.1678C3.19025 9.96143 2.87421 9.94284 2.66782 10.1263C2.46143 10.3098 2.44284 10.6258 2.6263 10.8322L6.6263 15.3322C6.81743 15.5472 7.15013 15.557 7.35356 15.3536L17.8536 4.85355C18.0488 4.65829 18.0488 4.34171 17.8536 4.14645C17.6583 3.95118 17.3417 3.95118 17.1465 4.14645L7.02141 14.2715L3.37371 10.1678Z")
+
+val cross20
+    get() = svgPath("M4.08859 4.21569L4.14645 4.14645C4.32001 3.97288 4.58944 3.9536 4.78431 4.08859L4.85355 4.14645L10 9.293L15.1464 4.14645C15.32 3.97288 15.5894 3.9536 15.7843 4.08859L15.8536 4.14645C16.0271 4.32001 16.0464 4.58944 15.9114 4.78431L15.8536 4.85355L10.707 10L15.8536 15.1464C16.0271 15.32 16.0464 15.5894 15.9114 15.7843L15.8536 15.8536C15.68 16.0271 15.4106 16.0464 15.2157 15.9114L15.1464 15.8536L10 10.707L4.85355 15.8536C4.67999 16.0271 4.41056 16.0464 4.21569 15.9114L4.14645 15.8536C3.97288 15.68 3.9536 15.4106 4.08859 15.2157L4.14645 15.1464L9.293 10L4.14645 4.85355C3.97288 4.67999 3.9536 4.41056 4.08859 4.21569L4.14645 4.14645L4.08859 4.21569Z")
+
+
+@JvmField
+val ERROR_CLASS: PseudoClass = PseudoClass.getPseudoClass("error")
+
+const val ERROR_STYLECLASS = "err"
+
+@Suppress("NOTHING_TO_INLINE")
+private inline fun svgPath(content: String) = SVGPath().also { it.content = content }
